use-cases/agent-factory-with-subagents/agents/pattern_integration_synthesizer/requirements.txt
# Pattern Integration Synthesizer Agent Dependencies

# Core framework
pydantic-ai>=0.0.14

# Data validation and settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# LLM providers
openai>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# HTTP client for MCP
httpx>=0.24.0
